Undo demo changes using steam
