+++ bevy_renet/Cargo.toml
@@ -13,18 +13,13 @@ version = "0.0.8"
 [features]
 default = ["transport"]
 transport = ["renet/transport"]
-steam_transport = ["renet_steam_transport", "steamworks"]
 
 [dependencies]
-
-
-bevy = { version = "0.11", default-features = false }
-renet = { path = "../renet", version = "0.0.12", features = ["bevy"] }
-renet_steam_transport = { path = "../renet_steam_transport", version = "0.0.1", features = [ "bevy" ], optional = true }
-steamworks = { git = "https://github.com/Noxime/steamworks-rs", rev = "2c77327", optional = true }
+bevy = {version = "0.11", default-features = false}
+renet = {path = "../renet", version = "0.0.12", features = ["bevy"]}
 
 [dev-dependencies]
 bevy = {version = "0.11", default-features = false, features = ["bevy_core_pipeline", "bevy_render", "bevy_asset", "bevy_pbr", "x11", "tonemapping_luts", "ktx2", "zstd"]}
 bincode = "1.3.1"
 env_logger = "0.10.0"
-serde = { version = "1.0", features = ["derive"] }
+serde = {version = "1.0", features = ["derive"]}
