+++ demo_bevy/Cargo.toml
@@ -16,11 +16,8 @@ bevy = {version = "0.11", default-features = false, features = ["bevy_core_pipel
 bevy_renet = { path = "../bevy_renet" }
 serde = { version = "1.0", features = [ "derive" ] }
 bincode = "1.3.1"
-
-
 # Use version directly when egui is updated to 0.22
 bevy_egui = "0.21.0"
 renet_visualizer = { path = "../renet_visualizer", features = ["bevy"] }
 smooth-bevy-cameras = { git = "https://github.com/bonsairobo/smooth-bevy-cameras.git", rev = "ed8ba36d3b202bc87bbae963670e1b9419804cd0" }
 fastrand = "2.0.0"
-steamworks = { git = "https://github.com/Noxime/steamworks-rs", rev = "2c77327" }
