+++ demo_chat/Cargo.toml
@@ -11,14 +11,7 @@ edition = "2021"
 renet = { path = "../renet" }
 renet_visualizer = { path = "../renet_visualizer" }
 eframe = "0.22"
-serde = { version = "1.0", features = ["derive"] }
+serde = { version = "1.0", features = [ "derive" ] } 
 bincode = "1.3.1"
-log = { version = "0.4.11", features = ["std"] }
+log = { version = "0.4.11", features = [ "std" ] }
 env_logger = "0.10.0"
-renet_steam_transport = { path = "../renet_steam_transport", version = "0.0.1", optional = true }
-steamworks = { git = "https://github.com/Noxime/steamworks-rs", rev = "2c77327", optional = true }
-
-[features]
-default = ["steam_transport"]
-transport = ["renet/transport"]
-steam_transport = ["renet_steam_transport", "steamworks"]
