+++ demo_chat/src/server.rs
@@ -1,77 +1,43 @@
-use std::{collections::HashMap, io, time::Duration};
+use std::{
+    collections::HashMap,
+    io,
+    net::UdpSocket,
+    time::{Duration, SystemTime},
+};
 
-#[cfg(feature = "transport")]
-use std::{net::UdpSocket, time::SystemTime};
-
-use renet::{ConnectionConfig, DefaultChannel, RenetServer, ServerEvent};
+use renet::{
+    transport::{NetcodeServerTransport, ServerAuthentication, ServerConfig},
+    ConnectionConfig, DefaultChannel, RenetServer, ServerEvent,
+};
 use renet_visualizer::RenetServerVisualizer;
 
-use crate::{client::PlatformState, ClientMessages, Message, ServerMessages};
-#[cfg(feature = "transport")]
-use crate::Username;
-
-#[cfg(feature = "transport")]
-use renet::transport::{NetcodeServerTransport, ServerAuthentication, ServerConfig};
-
-#[cfg(feature = "steam_transport")]
-use renet_steam_transport::server::{SteamServerTransport, SteamTransportConfig};
-
-#[cfg(feature = "transport")]
-use crate::PROTOCOL_ID;
-
+use crate::{ClientMessages, Message, ServerMessages, Username, PROTOCOL_ID};
 use bincode::Options;
 use log::info;
 
-pub enum ServerTransport {
-    #[cfg(feature = "transport")]
-    Netcode(NetcodeServerTransport),
-    #[cfg(feature = "steam_transport")]
-    Steam(SteamServerTransport),
-}
-
 pub struct ChatServer {
     pub server: RenetServer,
-    pub transport: ServerTransport,
+    pub transport: NetcodeServerTransport,
     pub usernames: HashMap<u64, String>,
     pub messages: Vec<Message>,
     pub visualizer: RenetServerVisualizer<240>,
 }
 
-impl ServerTransport {
-    pub fn new(platform: &PlatformState) -> Self {
-        match platform {
-            #[cfg(feature = "transport")]
-            PlatformState::Native => {
-                let socket = UdpSocket::bind("127.0.0.1:0").unwrap();
-                let public_addr = socket.local_addr().unwrap();
-                let server_config = ServerConfig {
-                    max_clients: 64,
-                    protocol_id: PROTOCOL_ID,
-                    public_addr,
-                    authentication: ServerAuthentication::Unsecure,
-                };
-
-                let current_time = SystemTime::now().duration_since(SystemTime::UNIX_EPOCH).unwrap();
-                let transport = NetcodeServerTransport::new(current_time, server_config, socket).unwrap();
-
-                ServerTransport::Netcode(transport)
-            }
-            #[cfg(feature = "steam_transport")]
-            PlatformState::Steam { steam_client, .. } => {
-                let config = SteamTransportConfig::from_max_clients(64);
-                let transport = SteamServerTransport::new(&steam_client, config).unwrap();
-
-                ServerTransport::Steam(transport)
-            }
-        }
-    }
-}
-
 impl ChatServer {
-    pub fn new(host_username: String, platform: &PlatformState) -> Self {
-        let server: RenetServer = RenetServer::new(ConnectionConfig::default());
+    pub fn new(host_username: String) -> Self {
+        let socket = UdpSocket::bind("127.0.0.1:0").unwrap();
+        let public_addr = socket.local_addr().unwrap();
+        let server_config = ServerConfig {
+            max_clients: 64,
+            protocol_id: PROTOCOL_ID,
+            public_addr,
+            authentication: ServerAuthentication::Unsecure,
+        };
 
-        let transport = ServerTransport::new(platform);
+        let current_time = SystemTime::now().duration_since(SystemTime::UNIX_EPOCH).unwrap();
+        let transport = NetcodeServerTransport::new(current_time, server_config, socket).unwrap();
+
+        let server: RenetServer = RenetServer::new(ConnectionConfig::default());
 
         let mut usernames = HashMap::new();
         usernames.insert(1, host_username);
@@ -87,28 +53,17 @@ impl ChatServer {
 
     pub fn update(&mut self, duration: Duration) -> Result<(), io::Error> {
         self.server.update(duration);
-        match &mut self.transport {
-            #[cfg(feature = "transport")]
-            ServerTransport::Netcode(netcode_transport) => {
-                netcode_transport.update(duration, &mut self.server).unwrap();
-            }
-            #[cfg(feature = "steam_transport")]
-            ServerTransport::Steam(steam_transport) => {
-                steam_transport.update(duration, &mut self.server);
-            }
-        }
+        self.transport.update(duration, &mut self.server).unwrap();
 
         self.visualizer.update(&self.server);
 
         while let Some(event) = self.server.get_event() {
             match event {
                 ServerEvent::ClientConnected { client_id } => {
-                    // let user_data = self.transport.user_data(client_id).unwrap();
-                    // let username = Username::from_user_data(&user_data).0;
-                    let username = "TODO".to_string();
-                    self.usernames.insert(client_id, username.clone());
-
+                    let user_data = self.transport.user_data(client_id).unwrap();
                     self.visualizer.add_client(client_id);
+                    let username = Username::from_user_data(&user_data).0;
+                    self.usernames.insert(client_id, username.clone());
                     let message = bincode::options()
                         .serialize(&ServerMessages::ClientConnected { client_id, username })
                         .unwrap();
@@ -141,16 +96,7 @@ impl ChatServer {
             }
         }
 
-        match &mut self.transport {
-            #[cfg(feature = "transport")]
-            ServerTransport::Netcode(netcode_transport) => {
-                netcode_transport.send_packets(&mut self.server);
-            }
-            #[cfg(feature = "steam_transport")]
-            ServerTransport::Steam(steam_transport) => {
-                steam_transport.send_packets(&mut self.server);
-            }
-        }
+        self.transport.send_packets(&mut self.server);
 
         Ok(())
     }
