--- demo_chat/steam_appid.txt
@@ -1,1 +1,0 @@
-480
